Add support for NTP to timestamp MQTT events.
When publishing to the the countPlaying and countIdle topics, if NTP is avialable we will use a timestamp for the current date/time instead of just saying "trigger".  I realized when I wanted to pull this into an InfluxDB that I actually needed timestamps to make the data meaningful.  If you don't care, then just undef the USE_NTP value near the start of the file.